Remove a util package; more cleanup for #86
